refactor: Move calculator logic to hook
